requirements.txt
# WikiStats — Python dependencies (Flask API + stream processing + Wikimedia producer)
# For stream_processor: Java 11+ and SPARK_HOME; Kafka connector jar via SPARK_KAFKA_PACKAGE or spark-submit --packages

Flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0

pymongo>=4.6.0
dnspython>=2.6.0

pyspark>=3.5.0,<3.6
kafka-python>=2.0.2
sseclient-py>=1.8.0
requests>=2.28.0

python-dotenv>=1.0.0
